% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alpha_MuSDrel.R
\name{alpha_muSDrel}
\alias{alpha_muSDrel}
\title{Parametrize shape parameter alpha for beta distribution}
\usage{
alpha_muSDrel(mu, sdrel)
}
\arguments{
\item{mu}{double - mean value within (0,1)}

\item{sdrel}{double - relative standard deviation}
}
\value{
double - The shape parameter alpha
}
\description{
This simple function computes alpha as a function of a relative standard deviation sdrel
and a mean value mu
}
\references{
Kruschke, John K. (2011). Doing Bayesian data analysis: A tutuorial with R and BUGS
}
