% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labels3d.R
\name{axLabels3d}
\alias{axLabels3d}
\title{Draw axis labels for 3d persp plots}
\usage{
axLabels3d(xlab = "x", ylab = "y", zlab = "z", cex = 2, pmat,
  dist = c(1.4, 1.3, 1.4), x, y, z)
}
\arguments{
\item{xlab}{String - X axis annotation. Defaults to "x".}

\item{ylab}{String - Y axis annotation. Defaults to "y".}

\item{zlab}{String - Z axis annotation. Defaults to "z".}

\item{cex}{double - character magnification. Defaults to 2.}

\item{pmat}{matrix of type double - A 4x4 matrix returned by calling graphics::persp}

\item{dist}{vector of type double - A three-dimensional vector containing the distances in (x,y,z)-direction between
axis and annotation. Defaults to c(1.4,1.3,1.4).}

\item{x}{vector of type double - The x range. It suffices, to supply c(min(x),max(x))}

\item{y}{vector of type double - The y range. It suffices, to supply c(min(y),max(y))}

\item{z}{vector of type double - The z range. It suffices, to supply c(min(z),max(z))}
}
\value{
No return value
}
\description{
This function draws axis labels for a 3d plot generated by graphics::persp
}
\details{
This function is especially designed for persp plots, since these lack manipulation options for
individual axis annotation texts. Please supply empty axis labels to persp, i.e. persp(xlab = "",ylab = "",zlab="",...).
}
\examples{
x = seq(from = -10,
to = 10,
length.out = 30)
y = x
f = function(x,y) {
  r =sqrt(x^2+y^2)
  10 * sin(r)/r
}
z = outer(x, y, f)
z[is.na(z)] = 1

par(mfrow=c(2, 1))

p1 = persp(x = x,
           y = y,
           z = z,
           theta = 30,
           phi = 30,
           expand = 0.5,
           col = "lightblue",
           ticktype="detailed",
           xlab="",
           ylab="",
           zlab="",
           r = 10)
myBayes::axLabels3d(x = x,
                    y = y,
                    z = c(min(z),max(z)),
                    pmat = p1,
                    xlab = "x label",
                    ylab = "y label",
                    zlab = "z label")

p1 = persp(x = x,
           y = y,
           z = z,
           theta = 30,
           phi = 30,
           expand = 0.5,
           col = "lightblue",
           ticktype="detailed",
           xlab="",
           ylab="",
           zlab="",
           r = 10)
myBayes::axLabels3d(x = x,
                    y = y,
                    z = c(min(z),max(z)),
                    pmat = p1,
                    dist = c(1.6,1.5,1.6),
                    xlab = "x label",
                    zlab = "z label")

}
\author{
Philipp van Wickevoort Crommelin
}
