% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta_muSDrel.R
\name{beta_muSDrel}
\alias{beta_muSDrel}
\title{Parametrize shape parameter beta for beta distribution}
\usage{
beta_muSDrel(mu, sdrel)
}
\arguments{
\item{mu}{double - mean value within (0,1)}

\item{sdrel}{double - relative standard deviation}
}
\value{
double - The shape parameter beta
}
\description{
This simple function computes beta as a function of a relative standard deviation sdrel
and a mean value mu
}
\references{
Kruschke, John K. (2011). Doing Bayesian data analysis: A tutuorial with R and BUGS
}
