% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/betaBin.R
\name{betaBin}
\alias{betaBin}
\title{Beta binomial distribution}
\usage{
betaBin(x, alpha, beta, n)
}
\arguments{
\item{x}{Double - A value within the interval [0,1].}

\item{alpha}{Double - A value greater than zero.}

\item{beta}{Double - A value greater than zero.}

\item{n}{Integer - Number of Bernoulli trials.}
}
\value{
Double - The corresponding probability.
}
\description{
The Beta Binomial distribution is a discrete probability distribution.
}
\author{
J.C. Lemm, P.v.W. Crommelin
}
