% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/betaDist.R
\name{betaDist}
\alias{betaDist}
\title{The Beta distribution}
\usage{
betaDist(x, alpha, beta)
}
\arguments{
\item{x}{Double - A value within the intervall [0,1].}

\item{alpha}{Double - A value greater than zero.}

\item{beta}{Double - A value greater than zero.}
}
\value{
Double - The corresponding probability.
}
\description{
The beta distribution is a continuous propability distribution defined in the interval [0,1].
}
\author{
J.C. Lemm, P.v.W. Crommelin
}
