% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blurrVector.R
\name{blurrVector}
\alias{blurrVector}
\title{blurrVector statistically disperses a deterministic vector}
\usage{
blurrVector(vec, lambda, sigmaRel = 0.1)
}
\arguments{
\item{vec}{Vector of type integer - A deterministic vector containing only zeros, except for one exposed component containing a one.}

\item{lambda}{A control parameter within [0,1].}

\item{sigmaRel}{Double - This value controlls the width of the maximum and eventually the dispersion of the resulting vector.}
}
\value{
A vector of type double - A probability vector, that is statistically dispersed around the exposed index position and normed to one.
}
\description{
This function is a convex combination of the Kronecker Delta function and the Betabinomial distribution.
}
\examples{
N = 150
deterministicVec = rep(0,N)
deterministicVec[N\%/\%4] = 1
probabilisticVec = blurrVector(vec = deterministicVec,
                               lambda = 0.1,
                               sigmaRel = 0.1)
plot(x = NULL,
     y = NULL,
     xlim = c(1,N),
     ylim = c(0,1),
     type = "p",
     cex = 3,
     col = "black",
     pch = 16)
M = 5
lambda = seq(from = 0,
             to = 1,
             length.out = M)
col = rainbow(M)
sapply(X = 1:M,
       FUN = function(i){
         points(x = 1:N,
               y = blurrVector(vec = deterministicVec,
                               lambda = lambda[i],
                               sigmaRel = 0.1),
               col = col[i])
       })
#' @export
}
\author{
P.v.W. Crommelin
}
