% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collapse.R
\name{collapse}
\alias{collapse}
\title{Remove equal and adjacent components of a vector}
\usage{
collapse(vec, order = TRUE)
}
\arguments{
\item{vec}{Vector of type double - The input vector.}

\item{order}{Boolean - Value indicating wether resulting dataframe should be sorted in decreasing order (defaults to TRUE).}
}
\value{
A data frame with two columns. 1st column $values: The resulting vector. 2nd column $indices:
The corresponding indices matching the input vector.
}
\description{
This functions takes a vector and removes duplicate, neighbouring entries. Indices are preserved under
this operation. If you encounter a sequence of equal values within a vector, the first one is preserved and its right
neighbours are removed.
}
\examples{
vec = c(1,1,3,4,5,6,7,7,7,7,7,8,9,10,9,8,7,6,5,4,4,4,4,3,2,1)
N = length(vec)
plot(x = 1:N,
     y = vec,
     type = "l",
     main = "demonstration")
points(x = 1:N,
       y = vec)
collapsedVec = collapse(vec = vec,
                        order = FALSE)
lines(x = collapsedVec$indices,
      y = collapsedVec$values,
      col = "red")
points(x = collapsedVec$indices,
       y = collapsedVec$values,
       col = "red")
legend("topright",
       leg = c("original vector","collapsed vector"),
       col = c("black","red"),
       lty = 1)
}
\author{
J.C. Lemm, P.v.W. Crommelin
}
