% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coordinates.R
\name{coordinates}
\alias{coordinates}
\title{Generate coordinates}
\usage{
coordinates(X, Y, surface)
}
\arguments{
\item{X}{Vector of type integer - x domain}

\item{Y}{Vector of type integer - y domain}
}
\value{
output is a matrix with 3 columns containing the coordinates X,Y,Z as column vectors
}
\description{
This function takes two vectors X and Y that define a domain and a surface Z,
and creates a matrix containing the coordinates.

X and Y are both 1D, surface is 2D
}
