% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/definiteness.R
\name{definiteness}
\alias{definiteness}
\title{Definiteness of a 2D real-valued matrix}
\usage{
definiteness(x)
}
\arguments{
\item{x}{matrix of type double - A real-valued, 2D matrix.}
}
\value{
character string - A character string indicating the definiteness of the given matrix.
}
\description{
This routine checks the definiteness of a two dimensional, real-valued matrix.
}
\details{
This routine takes the symmetric part of the given matrix and
decomposes its spectrum to find an answer.
}
\examples{
# positive definite
A = matrix(data = c(1,0,0,0,7,0,0,0,2),
           ncol = 3)
definiteness(A)

# negative definite
definiteness(-A)

# positive semi-definite
A = matrix(data = c(1,0,0,0,0,0,0,0,2),
           ncol = 3)
definiteness(A)

# negative semi-definite
definiteness(-A)

# Example, that demonstrates the importance of symmetry of the given matrix:
B = matrix(data = c(4,1,9,4),
           ncol = 2) #Eigenvalues all positive but not symmetric
definiteness(B)
eigen(B)$values
# Scalar product maps to negative value
v = c(-1,1)
t(v)\%*\%B\%*\%v

}
\author{
Philipp van Wickevoort Crommelin
}
