% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discreteMap_iter.R
\name{discreteMap_iter}
\alias{discreteMap_iter}
\title{Time Series generation from a matrix}
\usage{
discreteMap_iter(N, x0, mat, skipFirst = TRUE)
}
\arguments{
\item{N}{integer - number of iterations/Length of time series}

\item{x0}{double - starting value}

\item{mat}{matrix of type double - The generating system}

\item{skipFirst}{boolean - Wether the resulting time series should contain the initial value x0 or not}
}
\value{
the time series
}
\description{
This function produces a time Series generated by a supplied matrix
}
