% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{discretize_cpp}
\alias{discretize_cpp}
\title{discretize a real value}
\usage{
discretize_cpp(val, N_discr, domain, method)
}
\arguments{
\item{val}{double - real value to be discretized}

\item{N_discr}{integer - desired discretization}

\item{domain}{vector of double - vector with two components containing the domain bounds}

\item{method}{integer - discretize with 1..."round" or 2...custom method}
}
\value{
double - discretized value
}
\description{
this routine discretizes a value corresponding to the desired routine
}
\details{
This routine is implemented in C++
}
\examples{
//DEFINITION
double discretize_cpp(double val,  int N_discr, Rcpp::NumericVector domain, int method){
  double lower = domain[0];
  double upper = domain[1];
  if(method == 1){
    N_discr--;
    return(round(val*N_discr)/N_discr);
  }else{
    return(vecToVal_cpp(valToVec_cpp(val,N_discr,domain),domain));
  }
}
}
\author{
J.C. Lemm, P. v.W. Crommelin
}
