% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{gensymMap}
\alias{gensymMap}
\title{General symmetric map}
\usage{
gensymMap(x, r, alpha)
}
\arguments{
\item{x}{input value for one iteration}

\item{r}{control parameter}

\item{alpha}{exponent}
}
\value{
The result of one single iteration
}
\description{
a generalization of a one dimensional map, which incorporates (among others) the lorenz-, logistic and cubeMap
}
\references{
S. Sprott, Chaos and Time-series analysis
}
\author{
J.C. Lemm, P.v.W. Crommelin
}
