% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{gsm_cpp}
\alias{gsm_cpp}
\title{General symmetric map}
\usage{
gsm_cpp(x, r, alpha)
}
\arguments{
\item{x}{input value for one iteration}

\item{r}{control parameter}

\item{alpha}{exponent}
}
\value{
The result of one single iteration/
}
\description{
a generalization of a one dimensional map, which incorporates (among others) the lorenz-, logistic and cubeMap
}
\details{
This routine is implemented in C++
}
\references{
S. Sprott, Chaos and Time-series analysis
}
\author{
J.C. Lemm, P.v.W. Crommelin
}
