% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iteratedMap.R
\name{iteratedMap}
\alias{iteratedMap}
\title{Time series creation by given map.}
\usage{
iteratedMap(n, x0, fun, skipFirst = TRUE)
}
\arguments{
\item{n}{integer - length of resulting time series.}

\item{x0}{double - initial condition.}

\item{fun}{function of one parameter - the corresponding map.}

\item{skipfirst}{logical - wether the initial condition is to be removed.}
}
\value{
vector of type double and length n - the resulting time series.
}
\description{
Time series creation by given map.
}
\details{
This routine takes a map and applies it iteratively.
}
\examples{
N = 12
x = seq(from = 0,
       to = 1,
       by = 1/(N-1))
y = stats::runif(N)
fun = stats::approxfun(x = x,
                      y = y,
                      method = "linear")
x0 = 0.2
par(mfrow = c(2,1))
plot(x = x,
    y = y,
    main = "map",
    type = "l",
    col = "lightblue",
    lwd = 7,
    ylab = "")
N_series = 30
tmser = myBayes::iteratedMap(n = N_series,
                   x = x0,
                   fun = fun,
                   skipFirst = FALSE)
plot(NULL,
    xlim = c(1,N_series),
    ylim = c(1,0),
    main = "Time series",
    xlab = "iteration",
    ylab = "")
lines(x = 1:N_series,
     y = tmser,
     col = "purple",
     lwd = 5,
     ylab = "")
points(x = 1:N_series,
      y = tmser,
      col = "orange",
      cex = 2,
      pch = 16)
}
\author{
Philipp van Wickevoort Crommelin
}
