% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iteratedMap_ioNoise.R
\name{iteratedMap_ioNoise}
\alias{iteratedMap_ioNoise}
\title{Time series with input and measurement noise}
\usage{
iteratedMap_ioNoise(x0, r, n, funct = function(x, r) myBayes::gsm_cpp(x =
  x0, r = r, alpha = 2), irfunct = identity, orfunct = identity)
}
\arguments{
\item{x0}{double - starting value of timeseries}

\item{r}{double - bifurcation parameter}

\item{n}{integer - length of time series}

\item{funct}{function of two scalar values - map}

\item{irfunct}{function of scalar value - function for input noise}

\item{orfunct}{function of scalar value - output noise function}
}
\value{
the noisy time series
}
\description{
Time series with input and measurement noise
}
\details{
Iterated map with input and output noise
}
\examples{
x0 = 0.7
r = 0.8
N = 100
y = iteratedMap_ioNoise(x0 = x0,
                        r = r,
                        n = N,
                        funct=function(x,r) myBayes::gsm_cpp(x,r,alpha = 1.7),
                        orfunct = function(u) min(1,max(0,rnorm(n = 1,
                                                                mean = u,
                                                                sd = 0.1))))
plot(x = 1:N,
     y = y,
     type = "l",
     ylim = c(0,1))

}
