% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lik.R
\name{Lik}
\alias{Lik}
\title{Gaussian likelihood for given time series data}
\usage{
Lik(X, Y, sigma)
}
\arguments{
\item{X}{Vector of type double - The given model vector.}

\item{Y}{Vector of type double - The given data.}

\item{sigma}{Double - The standard deviation of the gaussian likelihood.}
}
\value{
Double - The resulting likelihood.
}
\description{
Lik computes the Gaussian likelihood for a the supplied data Y given the model vector X.
The standart deviation sigma is
equal for every datapoint in this implementation. Hence, please supply a single scalar value of type double.
}
\author{
Philipp van Wickevoot Crommelin
}
