% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Lik_gsm_cpp}
\alias{Lik_gsm_cpp}
\title{Gaussian likelihood for given data}
\usage{
Lik_gsm_cpp(alpha, r, x0, Y, sigma, N_discr)
}
\arguments{
\item{alpha}{double - exponent of gsm}

\item{r}{double - control parameter of gsm}

\item{x0}{double - initial value of the time series}

\item{Y}{vector of type double - the given data}

\item{sigma}{double - the standard deviation of the gaussian likelihood}

\item{N_discr}{integer - discretization of the state space (N_discr = 0 -> continuous case)}
}
\description{
Generates a time Series generated by the general symmetric map (gsm) and computes the likelihood given the data
}
\details{
This routine is implemented in C++
}
\references{
S. Sprott, Chaos and Time-series analysis
}
\author{
J.C. Lemm, P. v.W. Crommelin
}
