% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MH.R
\name{MH}
\alias{MH}
\title{Metropolis Hastings Algorithm}
\usage{
MH(T = 1, x0, trKern = dnorm, sampler, N = 1000, bounds)
}
\arguments{
\item{T}{double - temperature}

\item{x0}{double - starting value}

\item{trKern}{function of two values - Transition kernel as a function from two double values}

\item{sampler}{sampling distribution}

\item{N}{integer - size of time series}

\item{bounds}{array of size two - bounds for the search in 1D}
}
\value{
The resulting Markov Chain as an array of size N
}
\description{
Metropolis Hastings Algoritm. Create a random variable, which is distributed according to a custom probability distribution.
}
\author{
P. van Wickevoort Crommelin
}
