% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mu_alphaBeta.R
\name{mu_alphaBeta}
\alias{mu_alphaBeta}
\title{Transform shape parameters of Beta distribution}
\usage{
mu_alphaBeta(alpha, beta)
}
\arguments{
\item{alpha}{double - shape parameter >0}

\item{beta}{double - shape paramter >0}
}
\value{
the resulting mean value
}
\description{
This function transforms the shape paramters alpha and beta
of a Beta distribution to a mean value.
}
\author{
Philipp van Wickevoort Crommelin
}
