% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myGrid.R
\name{myGrid}
\alias{myGrid}
\title{myGrid draws grids}
\usage{
myGrid(xlim, ylim, nx, ny = nx, lty = DASHED, lwd = 1,
  col = "black")
}
\arguments{
\item{xlim}{Vector of type double - The bounds along the x axis for the grid.}

\item{ylim}{Vector of type double - The bounds along the y axis for the grid.}

\item{nx}{Integer - Number of vertical lines to be drawn.}

\item{ny}{Integer - Number of horizontal lines to be drawn. Defaults to nx.}

\item{lty}{Integer - Specifies the linetype similarly to R's core function grid().}

\item{lwd}{Integer - Specifies the linewidth similarly to R's core function grid().}

\item{col}{Character string - The color to be used for the grid lines.}
}
\value{
No return value.
}
\description{
Extension of R's core function grid(). This function plots a grid only within the box constraints specified
by xlim and ylim. The grid will be drawn to the currently open graphics device.
}
\examples{
x = seq(from = 0,
             to = 3,
             by = 0.01)
plot(x = x,
     y = x,
     type = "l")
myGrid(xlim = c(1,2),
       ylim = c(0.5,2.5),
       nx = 10,
       ny = 15,
       lty = 2,
       lwd = 2,
       col = "blue")
}
\author{
P.v.W. Crommelin
}
