% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ravrg}
\alias{ravrg}
\title{Running average}
\usage{
ravrg(v)
}
\arguments{
\item{v}{Vector of type double - data vector}
}
\value{
Vector of type double - The running mean/average
}
\description{
This function calculates the running average of a given data vector
}
\author{
Philipp van Wickevoort Crommelin
}
