% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sig_alphaBeta.R
\name{sig_alphaBeta}
\alias{sig_alphaBeta}
\title{Transform shape parameters of Beta distribution}
\usage{
sig_alphaBeta(alpha, beta)
}
\arguments{
\item{alpha}{double - shape parameter >0}

\item{beta}{double - shape paramter >0}
}
\value{
the resulting "relative standard deviation"
}
\description{
This function transforms the shape paramters alpha and beta
of a Beta distribution to a "relative standard deviation"
}
\author{
Philipp van Wickevoort Crommelin
}
