% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vecToVal.R
\name{vecToVal}
\alias{vecToVal}
\title{Projection of a vector onto a given axis}
\usage{
vecToVal(vec, x_lower = 0, x_upper = 1)
}
\arguments{
\item{vec}{vector of type double}

\item{x_lower}{double - lower bound}

\item{x_upper}{double - upper bound}
}
\value{
the inner product of the given vector vec and the given axis.
}
\description{
Projection of a vector onto a given axis
}
\author{
Philipp van Wickevoort Crommelin
}
