% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coordinates.R
\name{coordinates}
\alias{coordinates}
\title{Generate coordinates}
\usage{
coordinates(X, Y, surface)
}
\arguments{
\item{X}{coordinates and creates a grid in the form of a vector}

\item{X}{coordinates and creates a grid in the form of a vector}
}
\value{
output is a matrix with 3 columns containing the coordinates X,Y,Z as column vectors
}
\description{
X and Y are both 1D, surface is 2D
}
