% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{gsm_cpp}
\alias{gsm_cpp}
\title{general symmetric map}
\usage{
gsm_cpp(x, r, alpha)
}
\arguments{
\item{x}{double - input value for one iteration}

\item{r}{double - control parameter}

\item{alpha}{double - exponent}
}
\value{
double - the result of one single iteration/
}
\description{
a generalization of a one dimensional map, which incorporates (among others) the lorenz-, logistic and cubeMap
}
\details{
This routine is implemented in C++
}
\examples{
//C++ C++ DEFINITION
double gsm_cpp(double x, double r, double alpha){
  return(r*(1-pow(2*abs(x-0.5),alpha)));
}
}
\references{
S. Sprott, Chaos and Time-series analysis
}
\author{
J.C. Lemm, P.v.W. Crommelin
}
