% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{gsm_iter_cpp}
\alias{gsm_iter_cpp}
\title{Time series creation}
\usage{
gsm_iter_cpp(N, x0, r, alpha, N_discr, skipFirst)
}
\arguments{
\item{N}{integer - Number of iterations}

\item{x0}{double - starting value}

\item{r}{double - controll parameter}

\item{alpha}{double - exponent of general symmetric map}

\item{N_discr}{integer - controlls discretization of state space (zero corresponds to continuous case)}

\item{skipFirst}{Boolean - If set to FALSE, the resulting time series contains the initial value x0}
}
\value{
vector of type double - the resulting time series
}
\description{
Create time series produced by general symmetric map
}
\details{
This routine is implemented in C++
}
\author{
J.C. Lemm, P. v.W. Crommelin
}
