% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iteratedMap.R
\name{iteratedMap}
\alias{iteratedMap}
\title{Time series creation}
\usage{
iteratedMap(x, r, n, funct)
}
\arguments{
\item{r}{double - bifurcation parameter}

\item{n}{integer - length of resulting time series}

\item{funct}{function of starting value x0 - the corresponding map}

\item{x0}{double - starting value}
}
\value{
vector of type double and length n, i.e. the resulting time series.
}
\description{
Time series creation
}
\details{
This routine takes a map an applies it iteratively
}
\examples{
r0 = 0.918999
x0 = 0.5
N = 100
y = iteratedMap(x = x0,
                r = r0,
                n = N,
                funct = function(x,r) myBayes::gsm_cpp(x,
                                                       r,
                                                       alpha = 2.0))
plot(x = 1:N,
     y = y,
     ylim = c(0,1),
     type = "l")
}
